tweak the menu animations, use the new layout on the index page
