+++ packages/mgilbride/src/components/blog/Shoemaker.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from 'react';
 import { CSSObject } from '@emotion/core';
 import { makeSpace, responsiveBreakpoints, makeSize } from '../../utils/design';
-import { Layout } from '../layout/Layout';
+import { Layout } from '../layout/Layout__old';
 
 const styleContainer: CSSObject = {
   display: 'grid',
