--- packages/mgilbride/src/components/layout/Layout2.tsx
@@ -1,29 +1,0 @@
-import React, { FC, ReactNode, useCallback, useState } from 'react';
-import { CSSObject } from '@emotion/core';
-import { Header } from './header/Header';
-import { ContentContainer } from './content/ContentContainer';
-import { NavMenu } from './nav/NavMenu';
-import { responsiveBreakpoints } from '../../utils/design';
-
-const styleContainer: CSSObject = {
-  position: 'absolute',
-  top: 0,
-  bottom: 0,
-  left: 0,
-  right: 0,
-  maxWidth: responsiveBreakpoints.desktop,
-  margin: 'auto',
-};
-
-export const Layout: FC<{ footer?: ReactNode }> = ({ children, footer }) => {
-  const [open, setOpen] = useState(false);
-  const toggleOpen = useCallback(() => setOpen(!open), [open]);
-
-  return (
-    <div css={styleContainer}>
-      <Header {...{ open, toggleOpen }} />
-      <NavMenu open={open} />
-      <ContentContainer footer={footer}>{children}</ContentContainer>
-    </div>
-  );
-};
