+++ packages/mgilbride/src/components/layout/nav/HomeLink.tsx
@@ -1,4 +1,4 @@
-import React, { FC } from 'react';
+import React, { FC, useRef } from 'react';
 import Link, { LinkProps } from 'next/link';
 import { animated, useSpring } from 'react-spring';
 import { CSSObject } from '@emotion/core';
@@ -10,15 +10,18 @@ export const HomeLink: FC<LinkProps & { open: boolean }> = ({
   href,
   children,
 }) => {
+  const ref = useRef<HTMLAnchorElement | null>(null);
   const activeStyle = useActiveStyle(href);
   const style: CSSObject = {
     textTransform: 'uppercase',
   };
   const spring = useSpring({
-    from: {
-      opacity: 0,
-      color: open ? palette.contrast() : palette.text(),
-    },
+    from: ref.current
+      ? {
+          opacity: 0,
+          color: open ? palette.contrast() : palette.text(),
+        }
+      : {},
     to: {
       opacity: 1,
       color: open ? palette.contrast() : palette.text(),
@@ -27,7 +30,7 @@ export const HomeLink: FC<LinkProps & { open: boolean }> = ({
   return (
     <Link href={href}>
       {/* eslint-disable-next-line jsx-a11y/anchor-is-valid */}
-      <animated.a css={{ ...style, ...activeStyle }} style={spring}>
+      <animated.a ref={ref} css={{ ...style, ...activeStyle }} style={spring}>
         {children}
       </animated.a>
     </Link>
