+++ packages/mgilbride/src/pages/index.tsx
@@ -1,17 +1,27 @@
 import React, { FC } from 'react';
 import { CSSObject } from '@emotion/core';
 import Link from 'next/link';
-import { makeSpace } from '../utils/design';
+import { makeSpace, palette } from '../utils/design';
 import { Layout } from '../components/layout/Layout';
 import { ImgTrace } from '../components/ImgTrace';
 
 const styleContainer: CSSObject = {
-  display: 'grid',
-  justifySelf: 'center',
-  alignSelf: 'center',
+  display: 'flex',
+  justifyContent: 'center',
+  alignItems: 'center',
+  height: '100%',
+};
+
+const styleContent: CSSObject = {
+  display: 'flex',
+  flexDirection: 'column',
   img: {
-    height: '50vh',
-    paddingBottom: makeSpace('sm'),
+    maxWidth: '45vh',
+    margin: 'auto',
+    padding: makeSpace('lg'),
+  },
+  h3: {
+    fontWeight: 'bold',
   },
   p: {
     textAlign: 'center',
@@ -19,6 +29,7 @@ const styleContainer: CSSObject = {
     margin: 0,
     a: {
       textDecoration: 'underline',
+      color: palette.text(),
       textTransform: 'initial',
     },
   },
@@ -27,19 +38,21 @@ const styleContainer: CSSObject = {
 const Home: FC = () => (
   <Layout>
     <div css={styleContainer}>
-      <h3>
-        <p>Yo</p>
-      </h3>
-      <ImgTrace path="profile_circle.png" alt="profile" />
-      <div>
-        <p>{`I'm Matt, from Philly`}</p>
-        <p>
-          I like writing <a href="https://github.com/matthewtgilbride">code</a>,
-          and dogs
-        </p>
-        <p>
-          check out my <Link href="/about">about</Link> page
-        </p>
+      <div css={styleContent}>
+        <h3>
+          <p>Yo</p>
+        </h3>
+        <ImgTrace path="profile_circle.png" alt="profile" />
+        <div>
+          <p>{`I'm Matt, from Philly`}</p>
+          <p>
+            I like writing{' '}
+            <a href="https://github.com/matthewtgilbride">code</a>, and dogs
+          </p>
+          <p>
+            check out my <Link href="/about">about</Link> page
+          </p>
+        </div>
       </div>
     </div>
   </Layout>
